Move components into individual modules
