+++ src/components2/Article/Image/index.ts
@@ -1,0 +1,11 @@
+import Image from './Image';
+import ImageFull from './ImageFull';
+import ImageMulti from './ImageMulti';
+import Subtitle from './Subtitle';
+
+export default {
+  Image,
+  ImageFull,
+  ImageMulti,
+  Subtitle
+};
